add typecast from int64 to float32 to hopefully avoid error

+++ codes/metrics.py
@@ -9,7 +9,7 @@ def dice_coef(y_true, y_pred, smooth=1.0):
 
     y_true_f = K.flatten(y_true)
     y_pred_f = K.flatten(y_pred)
-    intersection = K.sum(y_true_f * y_pred_f)
+    intersection = K.sum(y_true_f.astype(np.float32) * y_pred_f.astype(np.float32)) # Change: Adding typecast to hopefully avoid error
     return (2. * intersection + smooth) / (
         K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
